+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.1.0 - 2023-08-29
+### Added
+* New welcome system:
+    * Members must wait 10 minutes before being confirmed
+    * Visual layout to select appropriate roles for the user
+
 ## 5.0.9 - 2023-06-11
 ### Added
 * [`ruff`](https://github.com/astral-sh/ruff) is now the main linter for the repository